requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.3
celery==5.5.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
clickhouse-driver==0.2.10
colorama==0.4.6
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-spectacular==0.27.2
gunicorn==21.2.0
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
psycopg-binary==3.2.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.14
